pyproject.toml
[tool.poetry]
name = "telegram-bot"
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
aiogram = "^3.8.0"
aiohttp = "^3.9.5"
pydantic-settings = "^2.3.3"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv"
]
# Include jupyter checks
extend-include = ["*.ipynb"]

line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "ANN",
    "RUF",
    "ASYNC"
]
ignore = ["ANN101"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
